add an offset to anchor links
When you're linking to a specific hash, it'd be useful that the link would transport you to the exact location of that location.

Because there's static header of `221px`, the offset here is `121px`. I'm not exactly sure how to calculate that, but some trial and error put me on this exact value. Note that when the header isn't collapsed this won't be right, but then you'd be linked to the top of the page, since it would be visible already.


for example in [Emission](https://github.com/artsy/emission) you link to the [JS Glossary]http://artsy.github.io/blog/2016/11/14/JS-Glossary() like `http://artsy.github.io/blog/2016/11/14/JS-Glossary/#yarn`.