Add support for the <g-emoji> html element
Support for rendering GitHub comment reactions was added with
https://github.com/artsy/artsy.github.io/pull/429. :tada:

The implementation followed GitHub's pattern of rendering emojis using a
system typeface by default and injecting an image tag with a fallback
image for platforms which may not have an emoji typeface pre-installed.

The custom `<g-emoji>` html element must be defined and registered to
the window before it can be used. This commit brings in the custom html
element as defined by GitHub in
https://github.com/github/g-emoji-element.