chore: remove traces of google site search
Google Custom Search must have been part of an earlier iteration of the
blog, because it appears to have been imported wholesale in the very
first commit:

https://github.com/artsy/artsy.github.io/commit/4f102804b9

Looks like some of this code hung around even after the later switch to
Lunr.

Might as well clean it up, since it appears to be dead code.